populate/lostnfound/pyproject.toml
[project]
name = "populate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
